# perf: get less proposal data on clone