--- a/apps/frontend/src/graphql/proposal/getProposalToClone.graphql
+++ b/apps/frontend/src/graphql/proposal/getProposalToClone.graphql
@@ -1,0 +1,9 @@
+query getProposalToClone($primaryKey: Int!) {
+  proposal(primaryKey: $primaryKey) {
+    primaryKey
+    questionary {
+      ...questionary
+      isCompleted
+    }
+  }
+}