Closes: https://github.com/UserOfficeProject/issue-tracker/issues/1245

## Description
This PR introduces a performance improvement by fetching less proposal data when cloning proposals.

Testing against STFC prod I get these timings. A small improvement as the questionary is still a large part of the proposal but does cut down on payload by about 35% for our 2025_1 round. 

| getProposal | getProposalToClone |
| ------------- | ------------- |
| 1.096276s | 0.422477s | 
| 0.479019s | 0.458030s |
| 0.420704s | 0.494965s |
| 0.443650s | 0.426792s |
| 0.668923s | 0.538003s |
| 0.601283s | 0.370382s |
| 1.082309s | 0.424525s |
| 1.045410s | 0.496527s |
| 0.457982s | 0.405587s |
| 0.756918s | 0.367396s |


## Motivation and Context
The change is required to improve the application's performance and responsiveness. Previously, when a proposal was cloned, all of its data was fetched, which was inefficient and slowed down the application.

## Changes
1. Modified the onClick function in ProposalTable.tsx to use the new getProposalToClone API instead of the getProposal API.
2. Added a new GraphQL query, getProposalToClone, which fetches only the necessary data needed to clone a proposal - the primaryKey and the questionary, improving the performance significantly.


## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Fixes Jira Issue

[https://jira.esss.lu.se/browse/](https://jira.esss.lu.se/browse/)

## Depends On

<!--- Does this PR depend on another PR that should be merged first or at the same time -->

## Tests included/Docs Updated?

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->

- [ ] I have added tests to cover my changes.
- [ ] All relevant doc has been updated